+++ Frontend/src/pages/Login/components/useLogin.ts
@@ -19,6 +19,7 @@ const useLogin = (setError: (msg: string) => void) => {
         headers: {
           "Content-Type": "application/json"
         },
+        credentials: "include",
         body: JSON.stringify({ email, password })
       })
