+++ Frontend/src/pages/Profile/components/Animation.tsx
